+++ test/support/org_file_mock.ex
@@ -7,4 +7,11 @@ defmodule Macroy.OrgFileMock do
  "* Education", "** Coding", "*** DONE Go play in an Elixir stream",
  "    CLOSED: [2019-08-18 Sun 10:50] DEADLINE: <2019-08-13 Tue> SCHEDULED: <2019-08-13 Tue>"]
   end
+
+  def prep_org_file_stream!("/entries_missing_todo_n_done.org") do
+    ["* Electronic Organization", "** Go over Google Assistant stuff in email",
+     "   SCHEDULED: <2019-08-06 Tue>",
+     "* Education", "** Coding", "*** Go play in an Elixir stream",
+     "    CLOSED: [2019-08-18 Sun 10:50] DEADLINE: <2019-08-13 Tue> SCHEDULED: <2019-08-13 Tue>"]
+  end
 end